*** empty log message ***
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@491 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ src/prim/simple_polygon_2d.inl
@@ -642,11 +642,11 @@ lass::io::MatlabOStream& operator<<(lass::io::MatlabOStream& oOStream,
 {
 	LASS_ENFORCE_STREAM(oOStream) << "lasthandle = patch(";
 	oOStream << "[" << iPolygon[0].x;
-	for (int i=1;i<iPolygon.size();++i)
+	for (size_t i=1;i<iPolygon.size();++i)
 		oOStream << "," << iPolygon[i].x;
 	oOStream << "],";
 	oOStream << "[" << iPolygon[0].y;
-	for (int i=1;i<iPolygon.size();++i)
+	for (size_t i=1;i<iPolygon.size();++i)
 		oOStream << "," << iPolygon[i].y;
 	oOStream << "],";
 	oOStream << "'Color'," << oOStream.color() << ");" << std::endl;
